Conflicts:
	implement/oglplus/detail/smart_values.ipp
	implement/oglplus/string/utf8.ipp
	include/oglplus/config/compiler.hpp
	include/oglplus/detail/boolean.hpp
	include/oglplus/error/basic.hpp
	include/oglplus/glsl_string.hpp
- Also updated the eagine submodule.

+++ implement/oglplus/buffer.ipp
@@ -46,7 +46,7 @@ _binding(BufferTarget target)
 		Error,
 		EnumParam(_binding_query(target))
 	);
-	return name;
+	return GLuint(name);
 }
 
 OGLPLUS_LIB_FUNC
@@ -60,7 +60,7 @@ _binding(BufferIndexedTarget target, GLuint index)
 		Error,
 		EnumParam(_binding_query(target))
 	);
-	return name;
+	return GLuint(name);
 }
 
 OGLPLUS_LIB_FUNC
